Splitting assembler into basic components (HMM, de Brujin graph, etc.)
